modularizing files, add directories, refactoring structure
